don't throw exception if file to hash doesn't exit
This fixes a bug found by Chukwuma; mem caches the dependencies of a memoized
function so that those dependencies can be re-hashed. However, prior to this
patch, if one of those dependencies was removed/renamed, an exception would be
raised rather then simply trying to rerun the function. The fix is obvious:
if the file doesn't exist when trying to hash, hash to a value that will cause
the cache lookup to fail (in this case we choose "NOT FOUND")