auto-include the file of build functions as dependencies
Each module in python has a .__file__ which is a string of the path
the module was loaded with. Each function gets a .__module__ which
is the string of which module the function was created in.

Earlier we couldn't use this as the decorators caused the functions
to appear to all be coming from mem's __init__.py or it's util.py.

Now we have the decorators copy over the original .__module__, so
we can now automatically have memoize use this info to depend on the
files that functions are coming from

Now stop blowing away your '.mem' directories ><