gcc: make_depends() can't be memoized
Currently, gcc.make_depends() is memoized, but the resulting deps
aren't set as deps of gcc.make_depends() so if any of them are
deleted, we still return the old list of dependencies and have
an error thrown when Node goes to check the file. This will force
a make_depends() rerun and the build will continue correctly.